@@ -1,6 +1,6 @@
 {
   "name": "mermaid-visual",
-  "version": "1.0.2.0",
+  "version": "1.1.0.0",
   "author": "Ilfat Galiev",
   "scripts": {
     "package": "webpack --config view-webpack.config",
@@ -26,7 +26,7 @@
     "d3-time-format": "4.1.0",
     "dompurify": "3.0.6",
     "handlebars": "^4.7.8",
-    "mermaid": "^10.9.1",
+    "mermaid": "11.1.1",
     "pegjs-loader": "^0.5.8",
     "powerbi-visuals-utils-dataviewutils": "6.0.1",
     "raw-loader": "^4.0.2",
@@ -70,7 +70,7 @@
     "ts-loader": "9.5.1",
     "ts-node": "10.9.1",
     "typescript": "5.3.2",
-    "webpack": "5.89.0",
+    "webpack": "^5.94.0",
     "webpack-bundle-analyzer": "4.4.2",
     "webpack-cli": "5.1.4",
     "webpack-dev-server": "4.15.1"
