@@ -157,6 +157,9 @@ const plugins = [
         window: 'realWindow',
         define: 'fakeDefine',
         powerbi: 'corePowerbiObject'
+    }),
+    new webpack.optimize.LimitChunkCountPlugin({
+        maxChunks: 1
     })
 ];
 
@@ -180,7 +183,7 @@ const devServer = {
 
 const optimization = {
     concatenateModules: true,
-    minimize: true // enable minimization for create *.pbiviz file less than 2 Mb, can be disabled for dev mode
+    minimize: false // enable minimization for create *.pbiviz file less than 2 Mb, can be disabled for dev mode
 };
 
 module.exports = {
